Extract metadata from pool helpers
