+++ src/components/contextual/pages/vebal/LMVoting/GaugesTable.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { PoolToken } from '@balancer-labs/sdk';
-import { computed, ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { useI18n } from 'vue-i18n';
 
@@ -14,7 +13,6 @@ import { getNetworkSlug } from '@/composables/useNetwork';
 import {
   isStableLike,
   isUnknownType,
-  poolMetadata,
   poolURLFor,
 } from '@/composables/usePoolHelpers';
 import { isSameAddress } from '@/lib/utils';
@@ -33,6 +31,7 @@ import IconLimit from '@/components/icons/IconLimit.vue';
 import { differenceInWeeks } from 'date-fns';
 import { oneSecondInMs } from '@/composables/useTime';
 import { buildNetworkIconURL } from '@/lib/utils/urls';
+import { poolMetadata } from '@/lib/config/metadata';
 
 /**
  * TYPES
@@ -252,12 +251,10 @@ function getPickedTokens(tokens: PoolToken[]) {
           <BalAssetSet :logoURIs="orderedTokenURIs(gauge)" :width="100" />
         </div>
       </template>
-      <template
-        #poolCompositionCell="{ pool, address, addedTimestamp, network }"
-      >
+      <template #poolCompositionCell="{ pool, address, addedTimestamp }">
         <div v-if="!isLoading" class="flex items-center py-4 px-6">
-          <div v-if="poolMetadata(pool.id, network)" class="text-left">
-            {{ poolMetadata(pool.id, network)?.name }}
+          <div v-if="poolMetadata(pool.id)" class="text-left">
+            {{ poolMetadata(pool.id)?.name }}
           </div>
           <TokenPills
             v-else
