+++ src/components/tables/BalClaimsTable.vue
@@ -1,5 +1,4 @@
 <script lang="ts" setup>
-import { computed, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router';
 
@@ -11,7 +10,6 @@ import useBreakpoints from '@/composables/useBreakpoints';
 import useNumbers, { FNumFormats } from '@/composables/useNumbers';
 import {
   isStableLike,
-  poolMetadata,
   orderedPoolTokens,
   orderedTokenAddresses,
 } from '@/composables/usePoolHelpers';
@@ -21,6 +19,7 @@ import { GaugePool } from '@/composables/useClaimsData';
 import { Gauge } from '@/services/balancer/gauges/types';
 import PoolWarningTooltip from '@/components/pool/PoolWarningTooltip.vue';
 import useNetwork from '@/composables/useNetwork';
+import { poolMetadata } from '@/lib/config/metadata';
 
 /**
  * TYPES
