+++ src/lib/config/metadata.ts
@@ -1,0 +1,9 @@
+import { POOLS } from '@/constants/pools';
+import { PoolMetadata } from '@/types/pools';
+
+/**
+ * Get metadata for a pool if it exists
+ */
+export function poolMetadata(id: string): PoolMetadata | undefined {
+  return POOLS.Metadata[id.toLowerCase()];
+}
